+++ enigma/__init__.py
@@ -1,0 +1,13 @@
+import os
+from dotenv import load_dotenv
+load_dotenv()
+
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.config["SECRET_KEY"] = os.getenv("SECRET_KEY")
+app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("SQLALCHEMY_DATABASE_URI")
+db = SQLAlchemy(app)
+
+from enigma import routes
