+++ enigma/routes.py
@@ -1,0 +1,47 @@
+from flask import render_template, url_for, flash, redirect
+from enigma import app
+from enigma.forms import RegistrationForm, LoginForm
+from enigma.models import User, Post
+
+posts = [
+    {
+        "author": "Korosium",
+        "title": "Blog Post 1",
+        "content": "First post content",
+        "date_posted": "April 20, 2018"
+    },
+    {
+        "author": "Korosium",
+        "title": "Blog Post 2",
+        "content": "First post content",
+        "date_posted": "April 20, 2018"
+    }
+]
+
+@app.route("/")
+@app.route("/home")
+def home():
+    return render_template('home.html', posts=posts)
+
+@app.route("/about")
+def about():
+    return render_template('about.html', title="About")
+
+@app.route("/register", methods=["GET", "POST"])
+def register():
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        flash(f"Account created for {form.username.data}!", "success")
+        return redirect(url_for("home"))
+    return render_template("register.html", title="Register", form=form)
+
+@app.route("/login", methods=["GET", "POST"])
+def login():
+    form = LoginForm()
+    if form.validate_on_submit():
+        if True:
+            flash("You have been logged in!", "success")
+            return redirect(url_for("home"))
+        else:
+            flash("Login unsuccessful. Please check email and password", "danger")
+    return render_template("login.html", title="Login", form=form)
